+++ synapse-client-rest/src/main/java/io/americanexpress/synapse/client/rest/client/BaseReactiveRestClient.java
@@ -1,0 +1,5 @@
+package io.americanexpress.synapse.client.rest.client;
+
+public class BaseReactiveRestClient {
+
+}
