+++ synapse-client-rest/src/main/java/io/americanexpress/synapse/client/rest/client/BaseRestClient.java
@@ -20,7 +20,6 @@ import io.americanexpress.synapse.client.rest.model.ClientHeaders;
 import io.americanexpress.synapse.client.rest.model.QueryParameter;
 import org.slf4j.ext.XLogger;
 import org.slf4j.ext.XLoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
@@ -32,14 +31,14 @@ import java.net.URI;
 import java.util.List;
 
 /**
- * <code>BaseRestClient</code> class is the parent for all REST clients.
+ * {@code BaseRestClient} class specifies the prototypes for all blocking REST clients.
  *
- * @param <I> input request responseType
- * @param <O> output response responseType
- * @param <H> httpHeadersFactory used for each backend service call
+ * @param <I> input request type
+ * @param <O> output response type
+ * @param <H> httpHeadersFactory used to set the HTTP headers for each web service call
  * @author Paolo Claudio
  */
-public abstract class BaseRestClient<I extends BaseClientRequest, O extends BaseClientResponse, H extends BaseClientHttpHeadersFactory<I>> {
+public abstract class BaseRestClient<I extends BaseClientRequest, O extends BaseClientResponse, H extends BaseClientHttpHeadersFactory<I>> extends BaseClient<I, O, H> {
 
     /**
      * Logger used for this client.
@@ -48,43 +47,15 @@ public abstract class BaseRestClient<I extends BaseClientRequest, O extends Base
 
     /**
      * Template for performing REST operations using default serialization options set in the BaseRestClientConfig.
-     * This can be overridden in child classes if different serialization options need to be set while connecting to backend services.
+     * This can be overridden in child classes if different serialization options need to be set while connecting to back end services.
      */
     protected RestTemplate restTemplate;
 
     /**
-     * URL of the backend service injected from the application properties.
-     */
-    protected String url;
-
-    /**
-     * HTTP method of the backend service.
-     */
-    protected HttpMethod httpMethod;
-
-    /**
-     * HTTP headers factory used to produce the custom HTTP headers required to consume the backend services.
-     */
-    @Autowired
-    protected H httpHeadersFactory;
-
-    /**
-     * Used to create query parameters for the URI.
-     */
-    @Autowired
-    private QueryParameterUriCreator queryParameterUriCreator;
-
-    /**
-     * Used to create path variables for the URI.
-     */
-    @Autowired
-    private PathVariableUriCreator pathVariableUriCreator;
-
-    /**
      * Default constructor creates an instance of BaseRestClient with default values.
      */
-    public BaseRestClient() {
-        setHttpMethod(HttpMethod.POST);
+    protected BaseRestClient() {
+        this.httpMethod = HttpMethod.POST;
     }
 
     /**
@@ -106,92 +77,32 @@ public abstract class BaseRestClient<I extends BaseClientRequest, O extends Base
     }
 
     /**
-     * Set the httpMethod.
-     *
-     * @param httpMethod the httpMethod to set
-     */
-    public void setHttpMethod(HttpMethod httpMethod) {
-        this.httpMethod = httpMethod;
-    }
-
-    /**
-     * Get the httpMethod.
-     *
-     * @return the httpMethod
-     */
-    public HttpMethod getHttpMethod() {
-        return httpMethod;
-    }
-
-    /**
-     * Get the url.
-     *
-     * @return the url
-     */
-    public String getUrl() {
-        return url;
-    }
-
-    /**
-     * Set the url.
-     *
-     * @param url the url to set
-     */
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-    /**
-     * Create the request entity used for the backend service.
-     *
-     * @param clientHeaders   headers for the backend service
-     * @param request         body of the request
-     * @param queryParameters parameters needed to be added to URI
-     * @param pathVariables   to add to the URI
-     * @return the request entity for the backend service
-     */
-    protected RequestEntity<I> createRequestEntity(ClientHeaders clientHeaders, I request, List<QueryParameter> queryParameters, String... pathVariables) {
-        StringBuilder urlBuilder = new StringBuilder(url);
-        urlBuilder.append(pathVariableUriCreator.createPathVariableUri(pathVariables));
-        urlBuilder.append(queryParameterUriCreator.createQueryParameterUri(queryParameters));
-        String updatedUrl = urlBuilder.toString();
-        URI uri = URI.create(updatedUrl);
-
-        // Create the HTTP headers for the service
-        HttpHeaders httpHeaders = httpHeadersFactory.create(clientHeaders, request, updatedUrl);
-
-        // Create the request
-        return new RequestEntity<>(request, httpHeaders, httpMethod, uri);
-    }
-
-    /**
      * Get the response from the service given the HTTP headers and request body.
      *
-     * @param clientHeaders      headers for the backend service
-     * @param request            body of the request
+     * @param clientHeaders      headers for the back end service
+     * @param clientRequest      body of the request
      * @param responseEntityType type of client response
      * @param pathVariables      variables needed to be added to URI
-     * @return the response body from the backend service
+     * @return the response body from the back end service
      */
-    public O callService(ClientHeaders clientHeaders, I request, Class<?> responseEntityType, String... pathVariables) {
-        return callService(clientHeaders, request, responseEntityType, null, pathVariables);
+    public O callService(ClientHeaders clientHeaders, I clientRequest, Class<?> responseEntityType, String... pathVariables) {
+        return callService(clientHeaders, clientRequest, responseEntityType, null, pathVariables);
     }
 
     /**
      * Get the response from the service given the HTTP headers and request body.
      *
-     * @param clientHeaders      headers for the backend service
-     * @param request            body of the request
+     * @param clientHeaders      headers for the back end service
+     * @param clientRequest      body of the request
      * @param responseEntityType type of client response
      * @param queryParameters    parameters needed to be added to URI
      * @param pathVariables      variables needed to be added to URI
-     * @return the response body from the backend service
+     * @return the response body from the back end service
      */
-
-    public O callService(ClientHeaders clientHeaders, I request, Class<?> responseEntityType, List<QueryParameter> queryParameters, String... pathVariables) {
+    public O callService(ClientHeaders clientHeaders, I clientRequest, Class<?> responseEntityType, List<QueryParameter> queryParameters, String... pathVariables) {
 
         // Create the request
-        RequestEntity<I> requestEntity = createRequestEntity(clientHeaders, request, queryParameters, pathVariables);
+        RequestEntity<I> requestEntity = createRequestEntity(clientHeaders, clientRequest, queryParameters, pathVariables);
 
         // Call the service
         ResponseEntity<?> responseEntity = restTemplate.exchange(requestEntity, responseEntityType);
@@ -205,33 +116,56 @@ public abstract class BaseRestClient<I extends BaseClientRequest, O extends Base
     /**
      * Get the response from the service given the HTTP headers and request body.
      *
-     * @param clientHeaders      headers for the backend service
-     * @param request            body of the request
+     * @param clientHeaders      headers for the back end service
+     * @param clientRequest      body of the request
      * @param responseEntityType type of client response
      * @param pathVariables      variables needed to be added to URI
-     * @return the response body from the backend service
+     * @return the response body from the back end service
      */
-    public List<O> callService(ClientHeaders clientHeaders, I request, ParameterizedTypeReference<List<O>> responseEntityType, String... pathVariables) {
-        return callService(clientHeaders, request, responseEntityType, null, pathVariables);
+    public List<O> callService(ClientHeaders clientHeaders, I clientRequest, ParameterizedTypeReference<List<O>> responseEntityType, String... pathVariables) {
+        return callService(clientHeaders, clientRequest, responseEntityType, null, pathVariables);
     }
 
     /**
      * Get the list of responses from the service given the HTTP headers and request body.
      *
-     * @param clientHeaders      headers for the backend service
-     * @param request            body of the request
+     * @param clientHeaders      headers for the back end service
+     * @param clientRequest      body of the request
      * @param responseEntityType type of client response
      * @param queryParameters    parameters needed to be added to URI
      * @param pathVariables      variables needed to be added to URI
-     * @return the response body from the backend service
+     * @return the response body from the back end service
      */
-    public List<O> callService(ClientHeaders clientHeaders, I request, ParameterizedTypeReference<List<O>> responseEntityType, List<QueryParameter> queryParameters, String... pathVariables) {
+    public List<O> callService(ClientHeaders clientHeaders, I clientRequest, ParameterizedTypeReference<List<O>> responseEntityType, List<QueryParameter> queryParameters, String... pathVariables) {
 
         // Create the request
-        RequestEntity<I> requestEntity = createRequestEntity(clientHeaders, request, queryParameters, pathVariables);
+        RequestEntity<I> requestEntity = createRequestEntity(clientHeaders, clientRequest, queryParameters, pathVariables);
 
         // Call the service
         ResponseEntity<List<O>> responseEntity = restTemplate.exchange(requestEntity, responseEntityType);
         return responseEntity.getBody();
+    }
+    
+    /**
+     * Create the request entity used for the back end service.
+     *
+     * @param clientHeaders   headers for the back end service
+     * @param request         body of the request
+     * @param queryParameters parameters needed to be added to URI
+     * @param pathVariables   to add to the URI
+     * @return the request entity for the back end service
+     */
+    private RequestEntity<I> createRequestEntity(ClientHeaders clientHeaders, I request, List<QueryParameter> queryParameters, String... pathVariables) {
+        StringBuilder urlBuilder = new StringBuilder(url);
+        urlBuilder.append(pathVariableUriCreator.createPathVariableUri(pathVariables));
+        urlBuilder.append(queryParameterUriCreator.createQueryParameterUri(queryParameters));
+        String updatedUrl = urlBuilder.toString();
+        URI uri = URI.create(updatedUrl);
+
+        // Create the HTTP headers for the service
+        HttpHeaders httpHeaders = httpHeadersFactory.create(clientHeaders, request, updatedUrl);
+
+        // Create the request
+        return new RequestEntity<>(request, httpHeaders, httpMethod, uri);
     }
 }
